+++ src/main/java/org/netcracker/educationcenter/elasticsearch/database/model/FTPFileObject.java
@@ -1,4 +1,4 @@
-package org.netcracker.educationcenter.elasticsearch.model;
+package org.netcracker.educationcenter.elasticsearch.database.model;
 
 import java.time.LocalDate;
 import java.util.UUID;
